Extract Class Declaration methods into classes + Simplify Use Statement Service
I finally got to move the individual functions from the class declaration service into their own classes, and now everything is fancy and uses dependency injection. I was also able to remove the dumb "use statement dto" concept which I feel would be confusing for new developers, and split the use statement service into two methods, albeit somewhat duplicated. At some point the construction of the class declaration service should be moved to the service locator via a factory.
